reorganize scope to have less code dupl. But I don't like it since it requires to import a trait. More stuff to try...
